Fixes on comments

+++ programs/cardinal-reward-distributor/src/instructions/claim_rewards.rs
@@ -59,7 +59,7 @@ pub fn handler<'key, 'accounts, 'remaining, 'info>(ctx: Context<'key, 'accounts,
             .unwrap()
             .checked_mul(reward_entry.multiplier as u128)
             .unwrap()
-            .checked_div((10 as u128).checked_pow(reward_distributor.multiplier_decimals as u32).unwrap())
+            .checked_div((10_u128).checked_pow(reward_distributor.multiplier_decimals as u32).unwrap())
             .unwrap();
 
         // if this will go over max supply give rewards up to max supply
